requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==1.6.1
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
python-dateutil==2.8.2
httpx==0.24.1
gotrue==1.3.0
postgrest==0.13.2
realtime==1.0.5
storage3==0.7.4
supabase==2.3.4
gunicorn==21.2.0
google-generativeai==0.3.2


